requirements.txt
openai-whisper
coqui-tts
requests 